requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
groq==0.9.0
pillow==10.4.0
requests==2.32.3
